requirements.txt
anyio==4.9.0
APScheduler==3.11.0
cachetools==4.2.2
certifi==2025.1.31
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-telegram-bot==22.1
pytz==2025.2
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
tornado==6.4.2
typing_extensions==4.14.0
tzdata==2025.1
tzlocal==5.3.1
urllib3==1.26.18